app navigation, implemented protected routes
