+++ client/src/context/index.ts
@@ -1,0 +1,2 @@
+export * from './AuthContext'
+export * from './ProtectedRoute'
