+++ client/src/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './Signin'
+export * from './Game'
